added maze, input keys, obstacles
