--- output/read.txt
@@ -1,1 +1,0 @@
-images
